Delete Job full implementation
Added no applications restriction. Remaining error message when trying to delete a Job with Applications

+++ src/main/java/acme/features/employer/job/EmployerJobDeleteService.java
@@ -1,9 +1,12 @@
 
 package acme.features.employer.job;
 
+import java.util.Collection;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import acme.entities.application.Application;
 import acme.entities.job.Job;
 import acme.entities.roles.Employer;
 import acme.framework.components.Errors;
@@ -71,11 +74,14 @@ public class EmployerJobDeleteService implements AbstractDeleteService<Employer,
 
 		Integer descriptorId = entity.getDescriptor().getId();
 
-		this.repository.deleteDuties(descriptorId);
-		this.repository.deleteDescriptor(descriptorId);
-		this.repository.deleteAudit(entity.getId());
-		this.repository.deleteAppli(entity.getId());
-		this.repository.delete(entity);
+		Collection<Application> linkedApps = this.repository.findManyApplicationsByJobId(entity.getId());
+		if (linkedApps.isEmpty()) {
+
+			this.repository.deleteDuties(descriptorId);
+			this.repository.deleteDescriptor(descriptorId);
+			this.repository.deleteAudit(entity.getId());
+			this.repository.delete(entity);
+		}
 
 	}
